pttc: avoid integer overflow warning
C-states are encoded as state - 1, wrapping around to 0xf encoding zero.

To avoid an integer overflow analyzer warning, cast the state value to
signed int before subtracting one, and casting it to 8b unsigned after
masking.

+++ pttc/src/parse.c
@@ -816,9 +816,9 @@ static int parse_c_state(uint8_t *state, uint8_t *sub_state, const char *input)
 		break;
 	}
 
-	*state = (uint8_t) ((maj - 1) & 0xf);
+	*state = (uint8_t) (((int) maj - 1) & 0xf);
 	if (sub_state)
-		*sub_state = (uint8_t) ((min - 1) & 0xf);
+		*sub_state = (uint8_t) (((int) min - 1) & 0xf);
 
 	return 0;
 }
